core(Master): Implement the controller for the master node
In the master file, the Flask application is created, and the configurations and the cached result are initialized. Here, the main routes are handled. In the update_config(), the requests are parsed in JSON, the list of workers is extracted from the data, and updates the global config with the new workers. The cache is cleared because all changes in configurations might invalidate the results already stored. For searching, if the query already appears in the cache, return it, otherwise, get the data, store it and rank it(alphabetical order used)