feat(Template): Implement template for the engine
The functionalities are implemented using a script, in which the submission of the inputs are handled. In the configuration of the workers, the URL and directory are sent to the local config, and then fetched as JSON. For the search table, the checked workers are pushed and the query is fetched. In the end, display the results found by generating a table.