config(Controller): Configure the trading system
This represents the main controller for this assignment. All of the events and instances are inserted and applied and the main functionalities for the acconts and orders are implemented. In the _match_orders function, the buy and sell orders are stored, and the best possible buys and sells are compared by their price. If the trade is possible, then a TradeExecuted event takes place, and the payment is done. The replay functions is used for the reconstruction from the event log